Allow migrations from single node etcd clusters
This change adds a post-start that will conditionally check the health
of the current member if the property `wait_on_member_health` is set to
true. This functionality is enabled by default.

This change also adds a check to update the peer url if necessary when
there is only one instance. At the moment, we are making the assumption
that you can only migrate to this etcd release in a single node
deployment.

[#155132619]

Signed-off-by: James Myers <jfmyers9@gmail.com>